working no softening

+++ src/pyInterface.h
@@ -3,6 +3,6 @@
 
 #include "trees.h"
 
-vector<body> basicRun(vector<body>&, vector<double> centre, vector<double> dim, int numIter, double dt);
+vector<body> basicRun(vector<body>&, vector<double> centre, vector<double> dim, __int64 numIter, double dt);
 
 #endif //PYINTERFACE
